WFLY-4255 Properly name classes servicing WildFlyContainerController functionality and improve JavaDoc
