+++ common/src/main/java/org/jboss/as/arquillian/container/controller/StopContainerWithTimeout.java
@@ -25,7 +25,7 @@ import org.jboss.arquillian.container.spi.Container;
 import org.jboss.arquillian.container.spi.event.ContainerControlEvent;
 
 /**
- * StopContainerWithTimeout coming from {@link WildFlyClientContainerController#stop(java.lang.String, int)}.
+ * {@link ContainerControlEvent} implementation fired in {@link ClientWildFlyContainerController#stop(java.lang.String, int)}.
  *
  * @author Radoslav Husar
  * @version Jan 2015
@@ -34,6 +34,10 @@ public class StopContainerWithTimeout extends ContainerControlEvent {
 
     private int timeout;
 
+    /**
+     * @param container container to stop
+     * @param timeout   graceful shutdown timeout in seconds
+     */
     public StopContainerWithTimeout(Container container, int timeout) {
         super(container);
         this.timeout = timeout;
