+++ common/src/main/java/org/jboss/as/arquillian/container/controller/command/StopWithTimeoutContainerCommand.java
@@ -24,7 +24,7 @@ package org.jboss.as.arquillian.container.controller.command;
 import org.jboss.arquillian.container.test.impl.client.deployment.command.AbstractCommand;
 
 /**
- * StopWithTimeoutContainerCommand coming from {@link org.jboss.as.arquillian.container.controller.WildFlyContainerContainerController}.
+ * StopWithTimeoutContainerCommand coming from {@link org.jboss.as.arquillian.container.controller.InContainerWildFlyContainerController}.
  *
  * @author Radoslav Husar
  * @version Jan 2015
